backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pymysql==1.1.1
cryptography==42.0.8
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.3.0
pydantic[email]==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.4
python-multipart==0.0.9
httpx==0.27.0
slowapi==0.1.9
python-dotenv==1.0.1
